refactor(web): isolate day calendar columns
